Allow calculating multiple percentiles
My use case was calculating percentiles for TLS and non-TLS data in the
same execution and without having to mess with the configuration. It's a
bit smarter now, relying on the string prefix, but otherwise remains
still dumb but good enough. Add some more docs and example in the sample

Change-Id: Ic67402218e266beff8b0a3e0c1d8d29927bf8530